move all file related methods from RODSConn to RC_file trait
